Move tracing to separate trace package
Remove extraneous logging
